Ecomerce_App/pyproject.toml
[project]
name = "ecomerce-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "markdown>=3.9",
    "pillow>=11.3.0",
    "requests>=2.32.5",
]
